Selection, Merge, and Integer Radix Sort
This originally started as a homework for a class on radix sort but first decided to add selection sort and merge sort. Not the most efficient way of writing them so I guess there is still more room for improvement. Most notably, radix sort was way more difficult than I expected. I'm still planning to add float radix sort in the future.